Add OverridesAsJson handling to orchestrator
