+++ src/protagonist/DLCS.Web/Configuration/ApplicationBuilderX.cs
@@ -1,9 +1,11 @@
 ﻿using System.Collections.Generic;
+using DLCS.Web.Response;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
+using Newtonsoft.Json;
 
 namespace DLCS.Web.Configuration;
 
@@ -75,4 +77,20 @@ public static class ApplicationBuilderX
         services.AddSingleton<IHttpMessageHandlerBuilderFilter, HeaderPropagationMessageHandlerBuilderFilter>();
         return services;
     }
+    
+    /// <summary>
+    /// Parse OverridesAsJson appSetting to strongly typed dictionary
+    /// </summary>
+    public static IServiceCollection HandlePathTemplates(this IServiceCollection services)
+        => services.PostConfigure<PathTemplateOptions>(opts =>
+        {
+            if (!string.IsNullOrEmpty(opts.OverridesAsJson))
+            {
+                var overridesDict = JsonConvert.DeserializeObject<Dictionary<string, string>>(opts.OverridesAsJson);
+                foreach (var (key, value) in overridesDict)
+                {
+                    opts.Overrides.Add(key, value);
+                }
+            }
+        });
 }
